Add Username checks

+++ QRhunter/app/src/main/java/com/example/qrhunter/AddPlayerActivity.java
@@ -1,5 +1,6 @@
 package com.example.qrhunter;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
@@ -7,14 +8,19 @@ import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.TextView;
 
 import com.google.android.gms.common.data.DataBufferObserverSet;
+import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 
 import java.util.ArrayList;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 /**
  * This class is used to edit player information
@@ -32,7 +38,7 @@ public class AddPlayerActivity extends AppCompatActivity {
         EditText emailEditText = findViewById(R.id.emailEditText);
         EditText numberEditText = findViewById(R.id.numberEditText);
         Button submitButton = findViewById(R.id.submitButton);
-
+        TextView errorText = findViewById(R.id.errorText);
 
 
         /*
@@ -50,42 +56,63 @@ public class AddPlayerActivity extends AppCompatActivity {
             // To do that we have to reconstruct the player object for deletion and addition
 
 
-            //TODO make sure new username is unique!
             //TODO maybe once add player can handle QR Codes, we might wanna revamp this
             Bundle bundle = getIntent().getExtras();
             String currentUserName = bundle.getString("username");
             String username = usernameEditText.getText().toString();
+            if(username.length() > 20 || username.length() <= 0 ){
+                errorText.setText("Username must be between 1-20 Characters");
+                return;
+            }
             String email = emailEditText.getText().toString();
             String number = numberEditText.getText().toString();
             Player newUser = new Player(username,email, Integer.parseInt(number), new ArrayList<>());
             Database db = new Database();
-            //TODO add on Failure listener
-            db.removePlayer(currentUserName);
-            db.addPlayer(newUser);
-            //TODO add on failure listener
-            //Need to delete the player from the qr too
-            db.getQrCodesFromPlayer(currentUserName)
-                            .addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
-                                @Override
-                                public void onSuccess(QuerySnapshot queryDocumentSnapshots) {
-                                    for(QueryDocumentSnapshot doc : queryDocumentSnapshots){
-                                        //TODO add failure listeners
-                                        //Not sure what they can do since no rollback but atleast we can
-                                        //let the user know there was a glitch
-                                        String hash = doc.getString("hash");
-                                        db.removePlayerFromQRCode(currentUserName, hash);
-                                        db.addScannedCode(new QRCode(hash ,null, null, 0),
-                                                newUser);
+            //making Sure it is unique
+            db.getPlayer(username).addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
+                @Override
+                public void onSuccess(DocumentSnapshot documentSnapshot) {
+                    if(documentSnapshot.exists())
+                    {
+                        //Username is not unique, so it is invalid
+
+                        errorText.setText("Username is taken dummmy");
+                    }
+                    else {
+                        //TODO add on Failure listener
+                        db.removePlayer(currentUserName);
+                        db.addPlayer(newUser);
+                        //TODO add on failure listener
+                        //Need to delete the player from the qr too
+                        db.getQrCodesFromPlayer(currentUserName)
+                                .addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
+                                    @Override
+                                    public void onSuccess(QuerySnapshot queryDocumentSnapshots) {
+                                        for(QueryDocumentSnapshot doc : queryDocumentSnapshots){
+                                            //TODO add failure listeners
+                                            //Not sure what they can do since no rollback but atleast we can
+                                            //let the user know there was a glitch
+                                            String hash = doc.getString("hash");
+                                            db.addScannedCode(new QRCode(hash ,null, null, 0),
+                                                    newUser);
+                                            db.removePlayerFromQRCode(currentUserName, hash);
+
+                                        }
+                                        SharedPreferences settings = getSharedPreferences("UserInfo", 0);
+                                        //Save this new username locally, how nice
+                                        SharedPreferences.Editor editor = settings.edit();
+                                        editor.clear();
+                                        editor.putString("Username", username);
+                                        editor.apply();
+                                        Intent intent = new Intent(AddPlayerActivity.this, MainActivity.class);
+                                        startActivity(intent);
                                     }
-                                    SharedPreferences settings = getSharedPreferences("UserInfo", 0);
-                                    //Save this new username locally, how nice
-                                    SharedPreferences.Editor editor = settings.edit();
-                                    editor.clear();
-                                    editor.putString("Username", username);
-                                    editor.apply();
-                                }
-                            });
-            finish();
+                                });
+                        //finish();
+                    }
+                }
+            });
+
         });
     }
 }
